dodawanie i usuwanie stadionu i klubu

+++ Football/Football/ViewModel/Window/MainViewModel.cs
@@ -10,7 +10,13 @@ namespace Football.ViewModel.Window
 {
     public class MainViewModel : BaseViewModel
     {
+        #region Service
+        StadiumService stadiumService;
+        ClubService clubService;
+        #endregion
+
         public ObservableCollection<StadiumViewModel> stadium { get; set; }
+        public ObservableCollection<ClubViewModel> club { get; set; }
         object _SelectedStadium;
         public object SelectedStadium
         {
@@ -29,49 +35,60 @@ namespace Football.ViewModel.Window
         }
         public RelayCommand AddStadiumCommand { get; set; }
         public RelayCommand RemoveStadiumCommand { get; set; }
+        public RelayCommand AddClubCommand { get; set; }
+        public RelayCommand RemoveClubCommand { get; set; }
 
         public MainViewModel()
         {
             InitializeCommands();
-            UpdateGrid();
+            UpdateStadiumGrid();
+            UpdateClubGrid();
         }
 
         private void InitializeCommands()
         {
             AddStadiumCommand = new RelayCommand(AddStadium);
             RemoveStadiumCommand = new RelayCommand(RemoveStadium);
+            AddClubCommand = new RelayCommand(AddClub);
+            RemoveClubCommand = new RelayCommand(RemoveClub);
 
             stadium = new ObservableCollection<StadiumViewModel>();
+            club = new ObservableCollection<ClubViewModel>();
+
+            stadiumService = new StadiumService();
+            clubService = new ClubService();
         }
+
+        #region Stadium
 
         void AddStadium(object parameter)
         {
             if (parameter == null) return;
             var values = (object[])parameter;
-            StadiumService stadiumService = new StadiumService();
             Stadium s = new Stadium
             {
                 name = values[0].ToString(),
                 city = values[1].ToString(),
-                country =values[2].ToString()
+                country = values[2].ToString()
             };
             stadiumService.AddStadium(s);
-            UpdateGrid();
+            UpdateStadiumGrid();
         }
 
         void RemoveStadium(object parameter)//usuwanie po wyszukiwaniu
         {
             if (parameter == null) return;
             var values = (StadiumViewModel)parameter;
-            StadiumService stadiumService = new StadiumService();
-            stadiumService.RemoveStadium(values.Name,values.City,values.Country);
-            UpdateGrid();
+            if (false)//validacja na tabele
+            {
+                stadiumService.RemoveStadium(values.Name, values.City, values.Country);
+                UpdateStadiumGrid();
+            }
 
         }
 
-        void UpdateGrid()//Bieda update , czysci grida i od nowa laduje
+        void UpdateStadiumGrid()//Bieda update , czysci grida i od nowa laduje
         {
-            StadiumService stadiumService = new StadiumService();
             var s = stadiumService.GetAllStadium();
             stadium.Clear();
             foreach (Stadium item in s)
@@ -79,5 +96,43 @@ namespace Football.ViewModel.Window
                 stadium.Add(new StadiumViewModel { Name = item.name, City = item.city, Country = item.country });
             }
         }
+
+        #endregion
+
+        #region Club
+
+        void AddClub(object parameter)
+        {
+            if (parameter == null) return;
+            var values = (object[])parameter;
+            if (true)
+            {
+
+                clubService.AddClub(values[0].ToString(), values[1].ToString());
+            }
+            else
+            {
+                //clubService.AddClub()
+            }
+            UpdateClubGrid();
+
+        }
+
+        void RemoveClub(object parameter)
+        {
+
+        }
+
+
+        void UpdateClubGrid()//Bieda update , czysci grida i od nowa laduje
+        {
+            var c = clubService.GetAllClub();
+            club.Clear();
+            foreach (Club item in c)
+            {
+                club.Add(new ClubViewModel { Name = item.name});
+            }
+        }
+        #endregion
     }
 }
